+++ src/server/configurations-manager.ts
@@ -1,5 +1,6 @@
 import { LoggerConfiguration } from '@enigmatis/polaris-logs';
-import { MiddlewareConfiguration } from '..';
+import { MiddlewareConfiguration, PolarisServerOptions } from '..';
+import { PolarisServerConfig } from '../config/polaris-server-config';
 
 export const getDefaultMiddlewareConfiguration = (): MiddlewareConfiguration => {
     return {
@@ -14,5 +15,23 @@ export const getDefaultLoggerConfiguration = (): LoggerConfiguration => {
         loggerLevel: 'info',
         writeToConsole: true,
         writeFullMessageToConsole: false,
+    };
+};
+
+export const getPolarisServerConfigFromOptions = (
+    options: PolarisServerOptions,
+): PolarisServerConfig => {
+    return {
+        ...options,
+        middlewareConfiguration:
+            options.middlewareConfiguration || getDefaultMiddlewareConfiguration(),
+        logger: options.logger || getDefaultLoggerConfiguration(),
+        applicationProperties: options.applicationProperties || { version: 'v1' },
+        allowSubscription: options.allowSubscription || false,
+        shouldAddWarningsToExtensions:
+            options.shouldAddWarningsToExtensions === undefined
+                ? true
+                : options.shouldAddWarningsToExtensions,
+        allowMandatoryHeaders: options.allowMandatoryHeaders || false,
     };
 };