+++ test/integration/server/test-server.ts
@@ -1,0 +1,62 @@
+import { RealitiesHolder } from '@enigmatis/polaris-common';
+import { ConnectionOptions, getPolarisConnectionManager } from '@enigmatis/polaris-typeorm';
+import { ExpressContext, PolarisServer, PolarisServerOptions } from '../../../src';
+import { TestContext } from './context/test-context';
+import { initConnection } from './dal/connection-manager';
+import * as polarisProperties from './resources/polaris-properties.json';
+import { resolvers } from './schema/resolvers';
+import { typeDefs } from './schema/type-defs';
+import { loggerConfig } from './utils/logger';
+
+export const connectionOptions: ConnectionOptions = {
+    type: 'postgres',
+    url: process.env.CONNECTION_STRING || '',
+    entities: [__dirname + '/dal/entities/*.ts'],
+    synchronize: true,
+    logging: true,
+};
+
+const customContext = (context: ExpressContext): Partial<TestContext> => {
+    const { req, connection } = context;
+    const headers = req ? req.headers : connection?.context;
+
+    return {
+        customField: 1000,
+        requestHeaders: {
+            customHeader: headers['custom-header'],
+        },
+    };
+};
+
+export async function startTestServer(
+    config?: Partial<PolarisServerOptions>,
+): Promise<PolarisServer> {
+    await initConnection(connectionOptions);
+    const options = { ...getDefaultTestServerConfig(), ...config };
+    const server = new PolarisServer(options);
+    await server.start();
+    return server;
+}
+
+export async function stopTestServer(server: PolarisServer): Promise<void> {
+    await server.stop();
+    if (getPolarisConnectionManager().connections.length > 0) {
+        await getPolarisConnectionManager()
+            .get()
+            .close();
+    }
+}
+
+const getDefaultTestServerConfig = (): PolarisServerOptions => {
+    return {
+        typeDefs,
+        resolvers,
+        customContext,
+        port: polarisProperties.port,
+        logger: loggerConfig,
+        supportedRealities: new RealitiesHolder(
+            new Map([[3, { id: 3, type: 'notreal3', name: 'default' }]]),
+        ),
+        connection: getPolarisConnectionManager().get(),
+    };
+};